pyproject.toml
[tool.poetry]
name = "seerealbox-core-backend"
version = "0.1.0"
description = ""
authors = ["Lee-GwangCheon <gclee@seerealbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
motor = "^3.4.0"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
